% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askgpt.R
\name{askgpt}
\alias{askgpt}
\title{askgpt}
\usage{
askgpt(prompt)
}
\arguments{
\item{prompt}{The prompt sent to chatGPT}
}
\value{
A character value with the response to be generated by ChatGPT.
}
\description{
askgpt
}
\examples{
\dontrun{
cat(ask_chatgpt("How many ducks does it take to screw in a lightulb?"))
}

}
